+++ trunk/src/FileLog.cpp
@@ -59,7 +59,7 @@ void NONS_LogStrings::init(const std::wstring &oldName,const std::wstring &newNa
 		return;
 	}
 	*str=0;
-	ulong entries=atoi((std::string)(char *)buffer);
+	ulong entries=atol((std::string)(char *)buffer);
 	ulong offset=str-buffer+1;
 	for (;buffer[offset]==10 || buffer[offset]==13;offset++);
 	bool newFormat=0;
